Fix for version upgrade
