--- .devcontainer/Dockerfile
@@ -1,1 +1,0 @@
-FROM node:16-buster
