--- .devcontainer/devcontainer.json
@@ -1,11 +1,0 @@
-{
-	"name": "Blog",
-	
-	"dockerComposeFile": [
-		"docker-compose.yml"
-	],
-
-	"service": "default",
-
-	"workspaceFolder": "/workspace",
-}
