--- .devcontainer/docker-compose.yml
@@ -1,23 +1,0 @@
-version: "3.8"
-
-services:
-  default:
-    build: .
-    volumes:
-      - ../:/workspace:cached
-      - ~/.aws/credentials:/root/.aws/credentials
-      - vscode-extension-nodejs:/root/.vscode-server/extensions
-    command: /bin/sh -c "while sleep 1000; do :; done;"
-    ports:
-      - "3000:3000"
-    networks:
-      - default
-
-volumes:
-  vscode-extension-nodejs:
-    external: true
-
-networks:
-  default:
-    external: true
-    name: development
