+++ components/articles/articleCard.tsx
@@ -1,5 +1,5 @@
 import { ArticleMetadata, ArticleTag } from '../../models/article';
-import Link from 'next/link';
+import NextLink from 'next/link';
 import { ARTICLES_PATH } from '../../shared/constant';
 import { Tag } from './tag-list'
 import { HeadingBlock } from '../textblock';
@@ -65,33 +65,34 @@ const ArticleCard: React.FC<Props> = ({articleMetadata, tags}) => {
   const paddingBottom = isSmallScreen ? 1 : 2;
 
   return (
-    <Link href={`${ARTICLES_PATH}/${articleMetadata["firstPublished"]}`}>
-      <Card sx={{ minWidth: 275}}>
-        <CardActionArea >
-          <CardContent sx={{paddingX: paddingX, paddingTop: paddingTop, paddingBottom: paddingBottom}}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              First Published: {getDateFromUnix(articleMetadata["firstPublished"], -5)} | Last Modified: {getDateFromUnix(articleMetadata["lastModified"], -5)} | Edition: {articleMetadata["edition"]}
+    <Card sx={{ minWidth: 275}} >
+      <CardActionArea 
+        component={NextLink}
+        href={`${ARTICLES_PATH}/${articleMetadata["firstPublished"]}`}
+        sx={{ textDecoration: 'none', color: 'inherit' }}>
+        <CardContent sx={{paddingX: paddingX, paddingTop: paddingTop, paddingBottom: paddingBottom}}>
+          <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
+            First Published: {getDateFromUnix(articleMetadata["firstPublished"], -5)} | Last Modified: {getDateFromUnix(articleMetadata["lastModified"], -5)} | Edition: {articleMetadata["edition"]}
+          </Typography>
+          {
+            (articleMetadata["series"] != null && articleMetadata["series"] != "-")  && 
+            <HeadingBlock size="h6" text={`《${articleMetadata["series"]}》系列`} />
+          }
+          <HeadingBlock size="h4" text={`【${articleMetadata["type"]}】${articleMetadata["title"]}`} />
+          {
+            (articleMetadata["subtitle"] != null && articleMetadata["subtitle"] != "-")  && 
+            <Typography variant="h5" sx={{ mb: 1.5 }} color="text.secondary">
+              {articleMetadata["subtitle"]}
             </Typography>
-            {
-              (articleMetadata["series"] != null && articleMetadata["series"] != "-")  && 
-              <HeadingBlock size="h6" text={`《${articleMetadata["series"]}》系列`} />
-            }
-            <HeadingBlock size="h4" text={`【${articleMetadata["type"]}】${articleMetadata["title"]}`} />
-            {
-              (articleMetadata["subtitle"] != null && articleMetadata["subtitle"] != "-")  && 
-              <Typography variant="h5" sx={{ mb: 1.5 }} color="text.secondary">
-                {articleMetadata["subtitle"]}
-              </Typography>
-            }
-            {
-              articleMetadata["tags"] != null && articleMetadata["tags"].map((tagName: string, index) => {
-                var foo = tags.filter((data: ArticleTag) => data["id"] == tagName)[0]
-                return <Tag key={index} articleTag={foo}/>
-              })
-            }
-          </CardContent>
-        </CardActionArea>
-      </Card>
-    </Link>
+          }
+          {
+            articleMetadata["tags"] != null && articleMetadata["tags"].map((tagName: string, index) => {
+              var foo = tags.filter((data: ArticleTag) => data["id"] == tagName)[0]
+              return <Tag key={index} articleTag={foo}/>
+            })
+          }
+        </CardContent>
+      </CardActionArea>
+    </Card>
   );
 }
