+++ components/nav/drawerItem.tsx
@@ -1,9 +1,7 @@
-import { ListItemText, ListItemButton, ListItemIcon, List} from '@mui/material';
-import { Box, Grid } from '@mui/material';
+import { ListItemText, ListItemButton, ListItemIcon} from '@mui/material';
 import { useAppSelector } from '../../shared/hooks';
-import Link from 'next/link';
+import NextLink from 'next/link';
 import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
-import { SvgIconComponent } from '@mui/icons-material'
 import { MILK } from '../../shared/constant';
 
 interface Props {
@@ -17,18 +15,16 @@ const DrawerItem: React.FC<Props> = ({name, icon, path, action}) => {
   const currentPath: string = useAppSelector(state => state["path"]["value"]);
 
   return (
-    <Link href={path}>
-      <ListItemButton onClick={action} >
-        <ListItemIcon sx={{ color: MILK }}>
-          {icon}
-        </ListItemIcon>
-        <ListItemText primary={name} primaryTypographyProps={{variant: "darkbutton"}}/>
-        {path == currentPath &&
-          <ListItemIcon sx={{ color: MILK,  marginRight: -5}}>
-            <ArrowBackIosIcon />
-          </ListItemIcon>}
-      </ListItemButton>
-    </Link>
+    <ListItemButton onClick={action} LinkComponent={NextLink} href={path} sx={{ textDecoration: 'none', color: 'inherit' }}>
+      <ListItemIcon sx={{ color: MILK }}>
+        {icon}
+      </ListItemIcon>
+      <ListItemText primary={name} primaryTypographyProps={{variant: "darkbutton"}}/>
+      {path == currentPath &&
+        <ListItemIcon sx={{ color: MILK,  marginRight: -5}}>
+          <ArrowBackIosIcon />
+        </ListItemIcon>}
+    </ListItemButton>
   );
 }
