+++ components/root.tsx
@@ -1,4 +1,3 @@
-import Grid from '@mui/material/Grid';
 import { Box } from '@mui/material';
 import { useLargeScreen, useSmallScreen } from '../functions/common';
 
@@ -20,21 +19,4 @@ export const PageContainer: React.FC<any> = (props: any) => {
       {props["children"]}
     </Box>
   )
-}
-
-export const Center: React.FC = (props: any) => {
-  return (
-    <Grid
-      container
-      spacing={0}
-      direction="column"
-      alignItems="center"
-      justifyContent="center"
-      style={{ minHeight: `calc(100% - 64px)` }}
-      >
-      <Box sx={{ width: { xs: '25%' } }}>
-        {props["children"]}
-      </Box>
-    </Grid>
-  );
 }
